Safari-609.1.11.1/LayoutTests/svg/dynamic-updates/SVGTextElement-dom-transform-attr.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGTextElement-dom-transform-attr.js
// [Expected rendering result] 'Test passed' message - and a series of PASS messages

description("Tests dynamic updates of the 'transform' attribute of the SVGTextElement object")
createSVGTestCase();

var textElement = createSVGElement("text");
textElement.setAttribute("x", "0");
textElement.setAttribute("y", "215");
textElement.textContent = "Test passed";

rootSVGElement.appendChild(textElement);
shouldBeNull("textElement.getAttribute('transform')", "");

function repaintTest() {
    textElement.setAttribute("transform", "translate(0,-200)");
    shouldBeEqualToString("textElement.getAttribute('transform')", "translate(0,-200)");

    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
